Amélioration gestion utilisateurs Modifications statistiques Ajout services admin

+++ src/main/java/org/sallaire/service/UserService.java
@@ -43,56 +43,107 @@ public class UserService implements UserDetailsService {
 		return new UserDto(user, true);
 	}
 
-	public void saveUser(String userName, String password, String role) {
-		LOGGER.info("Saving user {} with password {} and role {}", userName, StringUtils.repeat("X", password.length()), role);
-		Role convertedRole = null;
-		List<Role> roles = null;
-		User user = userRepository.findUserByName(userName);
-		if (user == null) {
-			user = new User(userName, passwordEncoder.encode(password));
-		} else {
-			user.setPassword(passwordEncoder.encode(password));
+	public void saveUser(String userName, String password, String role, UserDto currentUser) {
+		if (!currentUser.getRoles().contains(Role.ADMIN) && !currentUser.getRoles().contains(Role.SYSADMIN) && !currentUser.getName().equals(userName)) {
+			LOGGER.warn("Current user has not enough credentials to modify users");
+			return;
 		}
-		if (role == null) {
-			if (CollectionUtils.isEmpty(user.getRoles())) {
+		List<Role> roles = null;
+		Role convertedRole = null;
+		User user = userRepository.findUserByName(userName);
+		if (!(currentUser.getRoles().contains(Role.ADMIN) || currentUser.getRoles().contains(Role.SYSADMIN))) {
+			if (user == null) {
+				LOGGER.warn("Current user has no enough credentials to create users role");
+				return;
+			} else if (!user.getName().equals(currentUser.getName())) {
+				LOGGER.warn("Current user has no enough credentials to create users role");
+				return;
+			}
+		}
+		if (role != null) {
+			try {
+				convertedRole = Role.valueOf(role);
+				convertedRole = checkRole(convertedRole, currentUser.getRoles());
+			} catch (IllegalArgumentException e) {
+				LOGGER.warn("Input role {} is not a valide one, should be one of {}", role, Role.values());
+			}
+
+		}
+		if (convertedRole == null) {
+			if (user == null || CollectionUtils.isEmpty(user.getRoles())) {
 				LOGGER.info("No role specified for user {}, it will be set to {} by default", userName, Role.USER);
 				convertedRole = Role.USER;
 			} else {
 				roles = user.getRoles();
 				LOGGER.debug("User already exists in db, conserve its roles : {}", roles);
 			}
-		} else {
-			try {
-				convertedRole = Role.valueOf(role);
-			} catch (IllegalArgumentException e) {
-				LOGGER.warn("Input role {} is not a valide one, should be one of {}, the user will be saved with default role {}", role, Role.values(), Role.USER);
-				convertedRole = Role.USER;
-			}
 		}
-		if (convertedRole != null) {
-			if (roles == null) {
-				switch (convertedRole) {
-				case USER:
-
-					roles = Lists.newArrayList(Role.USER);
-					break;
-				case ADMIN:
+		if (roles == null) {
+			switch (convertedRole) {
+			case USER:
+				roles = Lists.newArrayList(Role.USER);
+				break;
+			case ADMIN:
+				if (currentUser.getRoles().contains(Role.ADMIN) || currentUser.getRoles().contains(Role.SYSADMIN)) {
 					roles = Lists.newArrayList(Role.USER, Role.ADMIN);
-					break;
-				case SYSADMIN:
-					roles = Lists.newArrayList(Role.USER, Role.ADMIN, Role.SYSADMIN);
-					break;
+				} else {
+					LOGGER.warn("Current user has no enough credentials to modify users role");
+					roles = Lists.newArrayList(Role.USER);
 				}
-				LOGGER.debug("Roles computed : {}", roles);
-				user.setRoles(roles);
+				break;
+			case SYSADMIN:
+				roles = Lists.newArrayList(Role.USER, Role.ADMIN, Role.SYSADMIN);
+				break;
 			}
-
-			userRepository.save(user);
-			LOGGER.info("User saved");
 		}
+		saveUser(userName, password, roles);
 	}
 
-	public UserDto getUser(String userName) {
+	private Role checkRole(Role wantedRole, List<Role> currentRoles) {
+		Role computedRole = null;
+		switch (wantedRole) {
+		case USER:
+			computedRole = Role.USER;
+			break;
+		case ADMIN:
+			if (currentRoles.contains(Role.ADMIN) || currentRoles.contains(Role.SYSADMIN)) {
+				computedRole = Role.ADMIN;
+			} else {
+				LOGGER.warn("Current user has no enough credentials to modify user role to {}", Role.ADMIN);
+				computedRole = null;
+			}
+			break;
+		case SYSADMIN:
+			if (currentRoles.contains(Role.SYSADMIN)) {
+				computedRole = Role.SYSADMIN;
+			} else {
+				LOGGER.warn("Current user has no enough credentials to modify user role to {}", Role.SYSADMIN);
+				computedRole = null;
+			}
+			break;
+		}
+		return computedRole;
+	}
+
+	private void saveUser(String userName, String password, List<Role> roles) {
+		LOGGER.info("Saving user {} with password {} and role {}", userName, StringUtils.repeat("X", password.length()), roles);
+
+		User user = userRepository.findUserByName(userName);
+		if (user == null) {
+			user = new User(userName, passwordEncoder.encode(password));
+		} else if (password != null) {
+			user.setPassword(passwordEncoder.encode(password));
+		}
+		user.setRoles(roles);
+		userRepository.save(user);
+		LOGGER.info("User saved");
+	}
+
+	public UserDto getUser(String userName, UserDto currentUser) {
+		if (!currentUser.getName().equals(userName) && currentUser.getRoles().contains(Role.ADMIN) || currentUser.getRoles().contains(Role.SYSADMIN)) {
+			LOGGER.warn("Current user has no enough credentials to get user details");
+			return null;
+		}
 		User user = userRepository.findUserByName(userName);
 		if (user != null) {
 			return new UserDto(user);
@@ -104,7 +155,11 @@ public class UserService implements UserDetailsService {
 		return userRepository.count() > 0;
 	}
 
-	public void deleteUser(String name) {
+	public void deleteUser(String name, UserDto currentUser) {
+		if (!currentUser.getName().equals(name) && currentUser.getRoles().contains(Role.ADMIN) || currentUser.getRoles().contains(Role.SYSADMIN)) {
+			LOGGER.warn("Current user has no enough credentials to delete user");
+			return;
+		}
 		User user = userRepository.findUserByName(name);
 		if (user != null) {
 			userRepository.delete(user);
@@ -128,7 +183,7 @@ public class UserService implements UserDetailsService {
 		LOGGER.info("Check existing user");
 		if (!hasUsers()) {
 			LOGGER.info("No user in db, creating default one");
-			saveUser("admin", "admin", Role.SYSADMIN.name());
+			saveUser("admin", "admin", Lists.newArrayList(Role.USER, Role.ADMIN, Role.SYSADMIN));
 		}
 	}
